Menambah foreign key dan meredirect superuser ke superuser.users

+++ database/migrations/2020_09_29_154723_create_role_user_table.php
@@ -15,8 +15,8 @@ class CreateRoleUserTable extends Migration
     {
         Schema::create('role_user', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id')->unsigned();
-            $table->integer('user_id_user')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
+            $table->bigInteger('user_id_user')->unsigned();
             $table->timestamps();
         });
     }
